Fixing bug where /v1/photos/list.json does not return the proper response. Closes #1086 #1053

+++ src/libraries/controllers/ApiPhotoController.php
@@ -491,7 +491,8 @@ class ApiPhotoController extends ApiBaseController
     if($allowDuplicate == '0')
     {
       $hashResp = $this->api->invoke("/{$this->apiVersion}/photos/list.json", EpiRoute::httpGet, array('_GET' => array('hash' => $attributes['hash'])));
-      if(!empty($hashResp['result']))
+      // the second condition is for backwards compatability between v2 and v1. See #1086
+      if(!empty($hashResp['result']) && $hashResp['result'][0]['totalRows'] > 0)
       {
         unlink($localFile);
         return $this->conflict('This photo already exists based on a sha1 hash. To allow duplicates pass in allowDuplicate=1', $hashResp['result'][0]);
